Reference

Reference

NTECARS.GasConditions — Type

GasConditions(;pressure::AbstractFloat, T_gas::AbstractFloat)

Contains the basis information of the pressure in Pa and translational temperature in K.

The pressure and translational temperature are used for calculating linewidths. When fitting, it should be remembered to update T_gas accordingly.

NTECARS.LaserConfiguration — Type

LaserConfiguration(;wavelength1::Float64, wavelength2::Float64, stokesspectrum::Spectrum) LaserConfiguration(;wavelength1::Float64, wavelength2::Float64, stokesspectrum::Spectrum, laser1profile = nothing, laser2profile = nothing)

Contains information on the lasers including central wavelengths and profiles.

wavelenghts_1 and wavelenghts_2 have to be provided in units of nm. The spectral range of the stokes_spectrum together with wavelenghts_1 and wavelenghts_2 determines the region in which the CARS spectrum is calculated.

Example

lasers = LaserConfiguration(
    wavelength_1    = 532e-9,
    laser_1_profile = Gaussian(0.2/2.35),
    wavelength_2    = 532e-9,
    laser_2_profile = Gaussian(0.2/2.35),
    stokes_spectrum = Spectrum([602.5e-9, 610e-9], [1.0,1.0], :wavelength)
)

NTECARS.average_to_detector_pixels — Method

Bins a finely resolved spectrum defined at wavelengths ν with intensity I to the wavelengths ν_pixel

NTECARS.intervals_around_points — Method

creates an array of intervals that consist of (start_point, end_point) for all points
in the positions array array with a width of window_width. Overlapping intervals ar merged

NTECARS.parameters_that_modify_distributions — Method

parameters_that_modify_distributions(sim_, parameter_update_function!, parameters)

Find out which parameters modify the distrution function of CO2 or N2 This is needed the know which paraeters can be extended with uncertainties via Measurements.jl since this is only supported for the distribution functions right now.

NTECARS.resample_to_resolution — Method

Samples the given spectrum at a given resolution via interpolation. Used for convolutions with
the instrumental broadening profile.

NTECARS.rovibrational_populations_database — Method

Generate vibrational and rovibrational populations for a distribution function. If any parameter has
an uncertainty using measurements.jl, then the results will have the corresponding uncertainties

NTECARS.rovibrational_populations_database — Method

Generate vibrational and rovibrational populations for a distribution function. If any parameter has
an uncertainty using measurements.jl, then the results will have the corresponding uncertainties

NTECARS.sampling_points_from_intervals — Method

creates an array of sampling points in the range specified by "limits".
In the specified intervals, the points are sampled with the fine resolution
Δx_fine and inbetween them with Δx_coarse

NTECARS.total_susceptibility — Method

total_susceptibility(sim::CARSSimulator)

test2